Move Jigsaw Inventory scripts to a Demo directory
