+++ Plugins/JigsawInventorySystem/Source/JigsawInventorySystem/JigsawInventorySystem.h
@@ -3,7 +3,6 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
 
 class FJigsawInventorySystemModule : public IModuleInterface
 {
