+++ Plugins/JigsawInventorySystem/Source/JigsawInventorySystem/Private/Demo/DemoInventoryComponent.cpp
@@ -1,9 +1,8 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "DemoInventoryComponent.h"
-
-#include "InventoryDemoItem.h"
+#include "Demo/DemoInventoryComponent.h"
+#include "Demo/InventoryDemoItem.h"
 
 
 // Sets default values for this component's properties
