+++ Plugins/JigsawInventorySystem/Source/JigsawInventorySystem/Private/Demo/InventoryDemoItem.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "InventoryDemoItem.h"
+#include "Demo/InventoryDemoItem.h"
 
 FInventoryDemoItem::FInventoryDemoItem()
 {
