+++ Plugins/JigsawInventorySystem/Source/JigsawInventorySystem/Public/Demo/DemoInventoryComponent.h
